create exceptions and failure

+++ lib/domain/use_cases/delete_todo_list_use_case.dart
@@ -1,15 +1,27 @@
+import '../../data/exceptions/repository_exceptions.dart';
+import '../failures/failure.dart';
 import '../repositories/todo_repository.dart';
 
 class DeleteTodoListUseCase {
   final TodoRepository repository;
 
   DeleteTodoListUseCase(this.repository);
-  Future<String?> call(int todoItemId) async {
+  Future<Failure?> call(int todoItemId) async {
+    if (todoItemId <= 0) {
+      return (Failure('Informe o id da tarefa'));
+    }
+
     try {
       await repository.deleteTodoItem(todoItemId);
       return null;
+    } on InvalidInputRepositoryException {
+      return (Failure('Parâmetros inválidos'));
+    } on DeleteRepositoryException {
+      return (Failure('Não foi possivel excluir a tarefa'));
+    } on RepositoryException catch (_) {
+      return (Failure('Falha na fonte de dados'));
     } catch (e) {
-      return 'Não foi possivel excluir a tarefa';
+      return (Failure('Falha inesperado'));
     }
   }
 }
